requirements.txt
# Flask and Core Dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Security and Performance
flask-talisman==1.1.0
flask-compress==1.14
flask-caching==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
